+++ index.d.ts
@@ -1,0 +1,5 @@
+export { IFieldConfig } from "./src/app/models/field-config.interface";
+export { NgxFormModule } from "./src/app";
+export { Field } from "./src/app/models/field.interface";
+export { DynamicFieldService } from "./src/app/services/dynamic-field.service";
+export { DynamicFormComponent } from "./src/app/containers/dynamic-form/dynamic-form.component";