+++ src/app.ts
@@ -1,22 +1,16 @@
 import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ReactiveFormsModule, FormsModule  } from '@angular/forms';
-import { QuillModule } from "ngx-quill";
-
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
+import { QuillModule } from 'ngx-quill';
 import { DynamicFieldDirective } from './app/components/dynamic-field/dynamic-field.directive';
 import { DynamicFormComponent } from './app/containers/dynamic-form/dynamic-form.component';
-import { FormInputComponent } from './app/components/form-input/form-input.component';
-import { FormSelectComponent } from './app/components/form-select/form-select.component';
-import { FormTextareaComponent } from './app/components/form-textarea/form-textarea.component'
-import { FormTextEditorComponent } from './app/components/form-text-editor/form-text-editor.component';
-import { FormInputHidden } from './app/components/form-hidden/form-hidden.component';
-import { FormUserComponent } from './app/components/form-user/form-user.component';
-import { FormRadioComponent } from "./app/components/form-radio/form-radio.component";
-import { FormCheckboxComponent } from "./app/components/form-checkbox/form-checkbox.component";
-import { TagInputModule } from "ngx-chips";
+import { DynamicPanelComponent } from './app/containers/dynamic-panel/dynamic-panel.component';
+import { FormNavModule } from './nav/nav-app';
+import { TagInputModule } from 'ngx-chips';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { DataService } from './app/services/data.service';
-import { DynamicFieldService } from "./app/services/dynamic-field.service";
+import { DynamicFieldService } from './app/services/dynamic-field.service';
+import { PreloadService, Components } from './app/services/preload.service';
+import { FieldConfigService } from './app/services/field-config.service';
 
 @NgModule({
     imports: [
@@ -25,36 +19,25 @@ import { DynamicFieldService } from "./app/services/dynamic-field.service";
         ReactiveFormsModule,
         FormsModule,
         QuillModule,
-        TagInputModule
+        TagInputModule,
+        FormNavModule
     ],
     declarations: [
-        FormInputHidden,
-        FormTextEditorComponent,
+        Components,
         DynamicFieldDirective,
         DynamicFormComponent,
-        FormInputComponent,
-        FormSelectComponent,
-        FormTextareaComponent,
-        FormUserComponent,
-        FormRadioComponent,
-        FormCheckboxComponent
+        DynamicPanelComponent
     ],
     exports: [
         DynamicFormComponent
     ],
     entryComponents: [
-        FormInputHidden,
-        FormInputComponent,
-        FormSelectComponent,
-        FormTextEditorComponent,
-        FormTextareaComponent,
-        FormUserComponent,
-        FormRadioComponent,
-        FormCheckboxComponent
+        Components
     ],
     providers: [
-        DataService,
-        DynamicFieldService
+        DynamicFieldService,
+        PreloadService,
+        FieldConfigService
     ],
     schemas: [
         NO_ERRORS_SCHEMA
@@ -62,5 +45,4 @@ import { DynamicFieldService } from "./app/services/dynamic-field.service";
 })
 export class NgxFormModule { }
 
-export {DynamicFieldService, DataService}
-
+export { DynamicFieldService };