+++ src/app/components/form-input/form-input.component.ts
@@ -1,18 +1,14 @@
-import { Component, ViewContainerRef } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { Field } from '../../models/field.interface';
-import { FieldConfig } from '../../models/field-config.interface';
+import { IFieldConfig } from '../../models/field-config.interface';
 
 @Component({
-  selector: 'form-input',
-  template: require('./form-input.component.html')
+    selector: 'form-input',
+    template: require('./form-input.component.html')
 })
-export class FormInputComponent implements Field {
-  field: FieldConfig;
-  group: FormGroup;
-  fields: FieldConfig[];
-
-  isShow () {
-    return !this.field.hidden;
-  }
+export class FormInputComponent implements Field  {
+    field: IFieldConfig;
+    group: FormGroup;
+    model: object;
 }