+++ src/app/components/form-select/form-select.component.ts
@@ -1,21 +1,14 @@
 import { Component } from '@angular/core';
 import { FormGroup } from '@angular/forms';
-
 import { Field } from '../../models/field.interface';
-import { FieldConfig } from '../../models/field-config.interface';
+import { IFieldConfig } from '../../models/field-config.interface';
 
 @Component({
-  selector: 'form-select',
-  templateUrl: './form-select.component.html'
+    selector: 'form-select',
+    template: require('./form-select.component.html')
 })
 export class FormSelectComponent implements Field {
-  field: FieldConfig;
-  group: FormGroup;
-  fields: FieldConfig[];
-
-
-  isShow () {
-    return !this.field.hidden;
-  }
-
+    field: IFieldConfig;
+    group: FormGroup;
+    model: object;
 }