+++ src/app/containers/dynamic-form/dynamic-form.component.ts
@@ -1,41 +1,54 @@
-import { Component, EventEmitter, Input, OnChanges, OnInit, Output } from '@angular/core';
-import { FormGroup, FormBuilder, Validators, FormArray} from '@angular/forms';
+import { Component, Input, OnChanges, OnInit } from '@angular/core';
+import { FormGroup, FormBuilder, Validators, FormControl } from '@angular/forms';
+import { IFieldConfig } from '../../models/field-config.interface';
+import { DynamicFieldService } from '../../services/dynamic-field.service';
+import { FieldConfigService } from '../../services/field-config.service';
 
-import { FieldConfig } from '../../models/field-config.interface';
-import { DataService } from '../../services/data.service';
+interface IFormConfig {
+    form: any;
+    fields: IFieldConfig[];
+}
 
 @Component({
     exportAs: 'dynamicForm',
     selector: 'dynamic-form',
-    template: require('./dynamic-form.component.html')
+    template: require('./dynamic-form.component.html'),
+    styles: [require('./dynamic-form.component.scss').toString()]
 })
 export class DynamicFormComponent implements OnChanges, OnInit {
-
-    @Input() config: FieldConfig[] = [];
+    @Input() formConfig: IFormConfig; // IFieldConfig[] = [];
     @Input() model: any;
     @Input() dataProvider: object;
+    @Input() lookups: object;
 
-    // @Output() submit: EventEmitter<any> = new EventEmitter<any>();
+    public form: FormGroup;
 
-    form: FormGroup;
-
-    get controls() { return this.config.filter(({ type }) => type !== 'button'); }
+    get controls() { return this.formConfig.fields.filter(({ type }) => type !== 'button'); }
     get changes() { return this.form.valueChanges; }
     get valid() { return this.form.valid; }
     get value() { return this.form.value; }
 
-    constructor(private fb: FormBuilder, private dataService: DataService) { }
+    constructor(private fb: FormBuilder, private dynamicFieldService: DynamicFieldService, private fieldConfigService: FieldConfigService) { }
 
-    ngOnInit() {
-        this.dataService.set(this.dataProvider);
-        this.form = this.createGroup();
+    public ngOnInit() {
+        this.form = this.fb.group({});
+        this.controls.forEach(control => {
+            this.form.addControl(control.name, this.createControl(control));
+        });
         if (this.model) {
             this.form.patchValue(this.model);
         }
-
+        this.formConfig.fields.forEach(field => {
+            if (field.lookup && this.lookups.hasOwnProperty(field.lookup)) {
+                field.options = this.lookups[field.lookup];
+                if (field.extract) { field.options = field.options.map(f => f[field.extract]); }
+            }
+        });
+        this.fieldConfigService.addFields(this.formConfig.fields);
+        // this.form.disable();
     }
 
-    ngOnChanges() {
+    public ngOnChanges() {
         if (this.form) {
             const controls = Object.keys(this.form.controls);
             const configControls = this.controls.map((item) => item.name);
@@ -47,58 +60,27 @@ export class DynamicFormComponent implements OnChanges, OnInit {
             configControls
                 .filter((control) => !controls.includes(control))
                 .forEach((name) => {
-                    const config = this.config.find((control) => control.name === name);
-                    this.form.addControl(name, this.createControl(config));
+                    const cfg = this.formConfig.fields.find((control) => control.name === name);
+                    this.form.addControl(name, this.createControl(cfg));
                 });
-
         }
     }
 
-  /**
-   * @description create FormGroup and FormControl for all general field, exclude custom field
-   * @return {FormGroup}
-   */
-    createGroup() {
-        const group = this.fb.group({});
-        this.controls.forEach(control => {
-          if (!control.custom)  group.addControl(control.name, this.createControl(control));
-        });
-        return group;
-    }
-
-    createControl(config: FieldConfig) {
-        const { disabled, required, minLength, maxLength, email, min, max, pattern, nullValidator, value } = config;
-        let validators = [];
-        if (required != undefined && required) { validators.push(Validators.required); }
-        if (minLength != undefined) { validators.push(Validators.minLength(minLength)); }
-        if (maxLength != undefined) { validators.push(Validators.maxLength(maxLength)); }
-        if (email != undefined) { validators.push(Validators.email); }
-        if (min != undefined) { validators.push(Validators.min(min)); }
-        if (max != undefined) { validators.push(Validators.max(max)); }
-        if (pattern != undefined) { validators.push(Validators.pattern(pattern)); }
-        if (nullValidator != undefined) { validators.push(Validators.nullValidator); }
+    public createControl(cfg: IFieldConfig): FormControl {
+        const { disabled, required, minLength, maxLength, email, min, max, pattern, value } = cfg;
+        const validators = [];
+        if (required !== undefined && required) { validators.push(Validators.required); }
+        if (minLength !== undefined) { validators.push(Validators.minLength(minLength)); }
+        if (maxLength !== undefined) { validators.push(Validators.maxLength(maxLength)); }
+        if (email !== undefined) { validators.push(Validators.email); }
+        if (min !== undefined) { validators.push(Validators.min(min)); }
+        if (max !== undefined) { validators.push(Validators.max(max)); }
+        if (pattern !== undefined) { validators.push(Validators.pattern(pattern)); }
 
         return this.fb.control({ disabled, value }, validators);
     }
 
-
-    setDisabled(name: string, disable: boolean) {
-        if (this.form.controls[name]) {
-            const method = disable ? 'disable' : 'enable';
-            this.form.controls[name][method]();
-            return;
-        }
-
-        this.config = this.config.map((item) => {
-            if (item.name === name) {
-                item.disabled = disable;
-            }
-            return item;
-        });
-    }
-
-    setValue(name: string, value: any) {
+    public setValue(name: string, value: any) {
         this.form.controls[name].setValue(value, { emitEvent: true });
     }
-
 }