+++ src/app/services/dynamic-field.service.ts
@@ -1,34 +1,21 @@
-import {
-  ComponentFactoryResolver,
-  Injectable,
-  Inject
-} from '@angular/core'
+import { Injectable, Type } from '@angular/core';
 
 @Injectable()
 export class DynamicFieldService {
 
-  factoryResolver;
-  rootViewContainer;
-  customComponents = {};
+    components = {};
 
-  constructor(@Inject(ComponentFactoryResolver) factoryResolver) {
-    this.factoryResolver = factoryResolver
-  }
-  setRootViewContainerRef(viewContainerRef) {
-    this.rootViewContainer = viewContainerRef
-  }
-  addDynamicComponent(DynamicComponent) {
-    const factory = this.factoryResolver.resolveComponentFactory(DynamicComponent);
-    const component = factory.create(this.rootViewContainer.parentInjector);
-    this.rootViewContainer.insert(component.hostView);
-    return component;
-  }
+    addField(name: string, component: any) {
+        if (!name || !name.length) { throw new Error('Failed to add new component. Type is incorrent, Type value: ' + name); }
+        if (!component) { throw new Error('Failed to add new component. Component is undefined'); }
+        this.components[name] = component;
+    }
 
-  storeCustomComponent(name, DynamicComponent) {
-    this.customComponents[name] = DynamicComponent;
-  }
-
-  getCustomComponent(name) {
-    return this.customComponents[name] || null;
-  }
+    getField(name: string) {
+        if (!this.components[name]) {
+            const supportedTypes = Object.keys(this.components).join(', ');
+            throw new Error(`Trying to use an unsupported field type "${name}". Supported types: ${supportedTypes}`);
+        }
+        return this.components[name];
+    }
 }