+++ CHANGELOG.md
@@ -1,3 +1,158 @@
+## [2.5.5](https://github.com/LabShare/ngx-forms/compare/v2.5.4...v2.5.5) (2018-12-19)
+
+
+### Bug Fixes
+
+* added disabled style for ngx-quill ([d5adeea](https://github.com/LabShare/ngx-forms/commit/d5adeea))
+* fixed label margin ([ac86b53](https://github.com/LabShare/ngx-forms/commit/ac86b53))
+* removed input types ([a1ed15c](https://github.com/LabShare/ngx-forms/commit/a1ed15c))
+
+## [2.5.4](https://github.com/LabShare/ngx-forms/compare/v2.5.3...v2.5.4) (2018-12-19)
+
+
+### Bug Fixes
+
+* **package:** update ngx-quill to version 4.0.0 ([20b23ed](https://github.com/LabShare/ngx-forms/commit/20b23ed))
+
+## [2.5.3](https://github.com/LabShare/ngx-forms/compare/v2.5.2...v2.5.3) (2018-12-10)
+
+
+### Bug Fixes
+
+* fixed input type for checkbox ([3021202](https://github.com/LabShare/ngx-forms/commit/3021202))
+
+## [2.5.2](https://github.com/LabShare/ngx-forms/compare/v2.5.1...v2.5.2) (2018-12-10)
+
+
+### Bug Fixes
+
+* fixed checkbox input control ([15696dc](https://github.com/LabShare/ngx-forms/commit/15696dc))
+* fixed Form nav does not reset after view change PPC-210 ([000b0e9](https://github.com/LabShare/ngx-forms/commit/000b0e9))
+* removed lodash ([cd7a2f4](https://github.com/LabShare/ngx-forms/commit/cd7a2f4))
+
+## [2.5.1](https://github.com/LabShare/ngx-forms/compare/v2.5.0...v2.5.1) (2018-11-27)
+
+
+### Bug Fixes
+
+* added bottom margin to each field ([af7f202](https://github.com/LabShare/ngx-forms/commit/af7f202))
+* added red asterix for required fields ([6c14c48](https://github.com/LabShare/ngx-forms/commit/6c14c48))
+* refactored fields config injection ([f65cf5f](https://github.com/LabShare/ngx-forms/commit/f65cf5f))
+
+# [2.5.0](https://github.com/LabShare/ngx-forms/compare/v2.4.1...v2.5.0) (2018-11-27)
+
+
+### Bug Fixes
+
+* added lint and lint fixes ([c37ec60](https://github.com/LabShare/ngx-forms/commit/c37ec60))
+* added tests ([8a83057](https://github.com/LabShare/ngx-forms/commit/8a83057))
+* fixed tests ([32e71cd](https://github.com/LabShare/ngx-forms/commit/32e71cd))
+* form style fixed ([e0687e3](https://github.com/LabShare/ngx-forms/commit/e0687e3))
+* removed stylelint ([d8a4e20](https://github.com/LabShare/ngx-forms/commit/d8a4e20))
+
+
+### Features
+
+* added ability to hide panel header if title is not specified ([f52323e](https://github.com/LabShare/ngx-forms/commit/f52323e))
+* added ability to permanently lock a panel PPC-151 ([39c5c61](https://github.com/LabShare/ngx-forms/commit/39c5c61))
+
+## [2.4.1](https://github.com/LabShare/ngx-forms/compare/v2.4.0...v2.4.1) (2018-11-21)
+
+
+### Bug Fixes
+
+* form does not show after update ([541e284](https://github.com/LabShare/ngx-forms/commit/541e284))
+
+# [2.4.0](https://github.com/LabShare/ngx-forms/compare/v2.3.3...v2.4.0) (2018-11-19)
+
+
+### Features
+
+* added left menu ([5a060e1](https://github.com/LabShare/ngx-forms/commit/5a060e1))
+
+## [2.3.3](https://github.com/LabShare/ngx-forms/compare/v2.3.2...v2.3.3) (2018-11-16)
+
+
+### Bug Fixes
+
+* PPC-143 fields and form config merged ([fa5ed4a](https://github.com/LabShare/ngx-forms/commit/fa5ed4a))
+* removed events ([3244608](https://github.com/LabShare/ngx-forms/commit/3244608))
+* removed ObserverService ([919c205](https://github.com/LabShare/ngx-forms/commit/919c205))
+* tests ([0cb4ee4](https://github.com/LabShare/ngx-forms/commit/0cb4ee4))
+
+## [2.3.2](https://github.com/LabShare/ngx-forms/compare/v2.3.1...v2.3.2) (2018-11-15)
+
+
+### Bug Fixes
+
+* added red asterix. Removed isShow ([672bba8](https://github.com/LabShare/ngx-forms/commit/672bba8))
+* fixed tests ([fcce0a3](https://github.com/LabShare/ngx-forms/commit/fcce0a3))
+
+## [2.3.1](https://github.com/LabShare/ngx-forms/compare/v2.3.0...v2.3.1) (2018-11-13)
+
+
+### Bug Fixes
+
+* added missing types ([cecaff8](https://github.com/LabShare/ngx-forms/commit/cecaff8))
+
+# [2.3.0](https://github.com/LabShare/ngx-forms/compare/v2.2.0...v2.3.0) (2018-11-07)
+
+
+### Bug Fixes
+
+* increased testing coverage ([2b17161](https://github.com/LabShare/ngx-forms/commit/2b17161))
+
+
+### Features
+
+* added ability to extract column from lookup object ([cc1d1bd](https://github.com/LabShare/ngx-forms/commit/cc1d1bd))
+
+# [2.2.0](https://github.com/LabShare/ngx-forms/compare/v2.1.1...v2.2.0) (2018-11-06)
+
+
+### Bug Fixes
+
+* **rjxs:** remove rxJS 5 import ([9b0c290](https://github.com/LabShare/ngx-forms/commit/9b0c290))
+* fixed tests ([9fcb982](https://github.com/LabShare/ngx-forms/commit/9fcb982))
+* fixed tests ([8fba67e](https://github.com/LabShare/ngx-forms/commit/8fba67e))
+* removed formComponentType from options ([bec44cf](https://github.com/LabShare/ngx-forms/commit/bec44cf))
+* removed test coverage assets ([16d93c7](https://github.com/LabShare/ngx-forms/commit/16d93c7))
+
+
+### Features
+
+* added lookup values ([452b017](https://github.com/LabShare/ngx-forms/commit/452b017))
+
+## [2.1.1](https://github.com/LabShare/ngx-forms/compare/v2.1.0...v2.1.1) (2018-08-14)
+
+
+### Bug Fixes
+
+* **test:** add unit test ([127831f](https://github.com/LabShare/ngx-forms/commit/127831f))
+* **test:** add unit test ([cd79259](https://github.com/LabShare/ngx-forms/commit/cd79259))
+* **test:** update ([e2b0892](https://github.com/LabShare/ngx-forms/commit/e2b0892))
+* **ui:** move out customized functionality ([144411e](https://github.com/LabShare/ngx-forms/commit/144411e))
+
+# [2.1.0](https://github.com/LabShare/ngx-forms/compare/v2.0.3...v2.1.0) (2018-08-07)
+
+
+### Bug Fixes
+
+* **refactor:** memory leak ([0a7f92b](https://github.com/LabShare/ngx-forms/commit/0a7f92b))
+* **typo:** update ([5857db5](https://github.com/LabShare/ngx-forms/commit/5857db5))
+
+
+### Features
+
+* **ui:** add dynamic panel #PPC-83 ([1300138](https://github.com/LabShare/ngx-forms/commit/1300138)), closes [#PPC-83](https://github.com/LabShare/ngx-forms/issues/PPC-83)
+
+## [2.0.3](https://github.com/LabShare/ngx-forms/compare/v2.0.2...v2.0.3) (2018-08-03)
+
+
+### Bug Fixes
+
+* **tests:** integrate with CodeCov ([ab364c5](https://github.com/LabShare/ngx-forms/commit/ab364c5))
+
 ## [2.0.2](https://github.com/LabShare/ngx-forms/compare/v2.0.1...v2.0.2) (2018-07-24)